requirements-win.txt
fastapi==0.110.0
uvicorn==0.29.0        
qrcode[pil]==7.4.2
pillow==10.3.0
python-multipart==0.0.9   
colorama==0.4.6           
